Updated PDF Generator
