+++ catalog-collections.json
@@ -375,18 +375,25 @@
       },
       {
         "authors": [
-          "Jeremy Schoeneman"
+          "Jeremy Schoeneman",
+          "Erik Daguerre"
         ],
         "classifiers": [
           "Plugin :: Client :: Email :: Attachment"
         ],
-        "description": "Generates a PDF file with a link which includes the campaign URL with the individual message_id required to track individual visits to a website. Visit https://github.com/y4utj4/pdf_generator for example template files to use with this plugin.",
+        "description": "Generates a PDF file from an html attachment that process client King Phisher Jinja variables allowing to embed links to your landing page so users that click the link in the PDF can be tracked when they visit.",
         "files": [
           {
-            "path-destination": "pdf_generator.py",
-            "path-source": "client/pdf_generator.py",
-            "signature": "APtrstsbA20r75huooSDlHv/JmJlAMErmalQZCmnFLrHkLz2sQPKHlAthEthm6sFo9/H7tizeY78qjQ/hO+5Me+lAH2Wxlp76RbKDHBB0o/ke68e9ItNzjx0U3IErYHCZ5BIDevTLlB2/Ukrezi2C5o2fW2+rCeWJ0c2KkzP2aVVmz9Z",
-            "signed-by": "github.com/zeroSteiner"
+            "path-destination": "pdf_generator/README.md",
+            "path-source": "client/pdf_generator/README.md",
+            "signature": "AQmThrHW/V+/M239tI+MF2yL3A+gQ15goH9rFwrZhxzncfYJCAOYa5+Fh0GcT6ulEtJMIiY9OEwgQkeATHgXxERAAbdKLcdEifxQJi4PtWB1J+jUIlF+Hff7df0ESdok2wcxReN2U+QPhM+4XuUsOwRDbpUVWf/hswAtvLB7i2/oNv4c",
+            "signed-by": "github.com/wolfthefallen"
+          },
+          {
+            "path-destination": "pdf_generator/__init__.py",
+            "path-source": "client/pdf_generator/__init__.py",
+            "signature": "AEIF/Eq0ykfN0E6iqCiteRUpI3soPUG8n51rG+IE7G+pOlzzuguu94/OATaHHiByeeuX9pORFnt2TLKhQIYhJeSsARM1GRgRq2iZfjxJozKrBnQrJKl0d6C5FHM8LsT+4ClS5eiN1IzByHuO+274RKmV2quUC26nR5QbyRMKBTK+2neq",
+            "signed-by": "github.com/wolfthefallen"
           }
         ],
         "homepage": "https://github.com/securestate/king-phisher-plugins",
@@ -394,11 +401,11 @@
         "requirements": {
           "minimum-version": "1.8.0",
           "packages": [
-            "reportlab"
+            "weasyprint==47"
           ]
         },
         "title": "Generate PDF",
-        "version": "1.1"
+        "version": "2.0"
       },
       {
         "authors": [
@@ -837,8 +844,8 @@
       }
     ]
   },
-  "created": "2019-05-01T13:53:48.759711+00:00",
+  "created": "2019-05-02T20:33:39.759946+00:00",
   "created-by": "github.com/wolfthefallen",
-  "signature": "ABEs+eYX0MhiyB6ChVzRyu+3X3euX+I0vrKTjK7TvQos8i/9j3HAQO+xDnrIFubvmrlBduntNAlYUusKEduVJGdCAJGWQ2+H/NDArk0Cjt8mFI/Eg60FfgzQK221d9NNixq/UbMyMSnUY0JIZX0S1jocVLaqEmwhpPKyv/KDy3GRWFn4",
+  "signature": "AMdy2HHyb6CmMgPLfaSBfALy6e8HuB28C/4EWAVfrlwPCUhN2WNHn3PRpUp2DgrvlAgPfdfuBIKVT3AHwBOjGRzXARsKkM8T/FEfbqQiMa4clhFV7U+slqbgmGU3YaHTWDoXn6IOMNfVo538NWqrQLgDPCeSh7VkAF1B7kz03GCZigaz",
   "signed-by": "github.com/wolfthefallen"
 }
